dlp/requirements.txt
google-cloud-dlp==3.1.1
google-cloud-storage==1.39.0
google-cloud-pubsub==2.6.0
google-cloud-datastore==2.1.3
google-cloud-bigquery==2.20.0
gsutil
